move transaction header
